Crowd Detection using nanonets/requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.1
opencv-python==4.1.2.30
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
requests==2.22.0
six==1.14.0
urllib3==1.25.8
